el switch magico

+++ score.h
@@ -8,19 +8,19 @@ private:
 	char nombre[30];
 	int rondas;
 	int tiros, tirosacertados;
-	float accuracy;
+	int accuracy;
 public:
 	///set
 	void set_rondas(int a) { rondas = a; }
 	void set_nombre(const char* v) { strcpy(nombre, v); }
-	void set_accuracy(float _accuracy) { accuracy = _accuracy; }
+	void set_accuracy(int _accuracy) { accuracy = _accuracy; }
 	void set_tiros(int a) { tiros = a; }
 	void set_tirosacertados(int a) { tirosacertados = a; }
 
 	///get
 	int get_rondas() { return rondas; }
 	const char* get_nombre() { return nombre; }
-	float get_accuracy() { return accuracy; }
+	int get_accuracy() { return accuracy; }
 	int get_tiros() { return tiros; }
 	int get_tirosacertados() { return tirosacertados; }
